+++ src/services/managedPrefixList/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddawsManagedPrefixListInput!]!) {
+    addawsManagedPrefixList(input: $input, upsert: true) {
+        numUids
+    }
+}`