+++ src/services/transitGatewayRouteTable/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddawsTransitGatewayRouteTableInput!]!) {
+    addawsTransitGatewayRouteTable(input: $input, upsert: true) {
+      numUids
+    }
+  }`